backend/requirements.txt
# Core FastAPI and web server
fastapi==0.116.1
uvicorn==0.34.0
python-multipart==0.0.19
starlette==0.41.3

# HTTP clients and async
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.9.0
sniffio==1.3.1

# Data validation and processing
pydantic==2.10.8
pydantic-core==2.27.6
typing_extensions==4.12.2

# PDF processing
PyMuPDF==1.24.15
PyMuPDFb==1.24.15
python-dotenv==1.0.1

# PDF report generation
reportlab==4.0.9
matplotlib==3.8.2
numpy==1.26.2

# Data handling
pandas==2.3.1
python-dateutil==2.9.0.post0

# Utilities
click==8.2.1
packaging==24.3
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
